requirements.txt
# Core dependencies
aiohttp>=3.9.0
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0

# Search APIs
google-search-results>=2.4.0

# Compression
brotli>=1.1.0

# Progress display
tqdm>=4.66.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
